+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Show artist name and track name with link to Jamendo website
+
 ## [0.2.1] - 2020-05-29
 
 ### Added